+++ src/hooks/gql.ts
@@ -76,12 +76,6 @@ export function useCurrentOrg() {
   }).data?.organizations_by_pk;
 }
 
-export function useCircleIdForEpoch(epochId: number) {
-  return useTypedQuery(`circle-for-epoch-${epochId}`, {
-    epochs_by_pk: [{ id: epochId }, { circle_id: true }],
-  }).data?.epochs_by_pk?.circle_id;
-}
-
 export function useCurrentCircleIntegrations() {
   const { circleId } = useSelectedCircle();
   const query = useTypedQuery(`circle-integrations-${circleId}`, {
